Reorganized files a bit, not final yet
