+++ mala/descriptors/symmetric_grp_manip.py
@@ -1,7 +1,9 @@
 import numpy as np
-from sympy.combinatorics import Permutation
-from mala.descriptors.tree_sorting import *
-import itertools
+
+from mala.descriptors.label_sublib.young import (
+    local_sigma_c_partitions,
+    flatten,
+)
 
 
 def leaf_filter(lperms):
@@ -118,3 +120,17 @@ def enforce_sorted_orbit(partition_inds):
     # that call this function. I'll comment it out for now, in case it is
     # important for later debugging.
     return part_tup  # , tuple(new_partition)
+
+
+def group_vec_by_orbits(vec, part):
+    ind_range = np.sum(part)
+    assert len(vec) == ind_range, "vector must be able to fit in the partion"
+    count = 0
+    by_orbits = []
+    for orbit in part:
+        orbit_vec = []
+        for i in range(orbit):
+            orbit_vec.append(vec[count])
+            count += 1
+        by_orbits.append(tuple(orbit_vec))
+    return tuple(by_orbits)
